refactor: share droidrun protocol parser
