+++ openclaw-plugin-mobile-ui/src/internal/droidrun/protocol.ts
@@ -1,0 +1,20 @@
+export function parseJsonFromStdout(stdout: string): any | null {
+  const trimmed = (stdout || "").trim();
+  if (!trimmed) return null;
+
+  try {
+    return JSON.parse(trimmed);
+  } catch {}
+
+  const lines = trimmed
+    .split(/\r?\n/)
+    .map((line) => line.trim())
+    .filter(Boolean);
+  for (let i = lines.length - 1; i >= 0; i -= 1) {
+    try {
+      return JSON.parse(lines[i]);
+    } catch {}
+  }
+
+  return null;
+}
